+++ Ecommerce/App.js
@@ -4,6 +4,7 @@ import Routes from "./src/routes";
 import { StatusBar } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 
+
 export default function App() {
   return(
     <NavigationContainer>
