+++ Ecommerce/package.json
@@ -20,11 +20,11 @@
     "react-dom": "18.1.0",
     "react-native": "0.70.5",
     "react-native-animatable": "^1.3.3",
+    "react-native-gesture-handler": "~2.8.0",
+    "react-native-reanimated": "~2.12.0",
     "react-native-safe-area-context": "4.4.1",
     "react-native-screens": "~3.18.0",
-    "react-native-web": "~0.18.9",
-    "react-native-gesture-handler": "~2.8.0",
-    "react-native-reanimated": "~2.12.0"
+    "react-native-web": "~0.18.9"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
