+++ Ecommerce/src/routes/index.js
@@ -1,48 +1,10 @@
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
 
-import { Welcome } from "../components/screens/Welcome";
-import { Login } from "../components/screens/Login";
-import { RecuperarSenha } from "../components/screens/RecuperarSenha";
-import SobreNos from "../components/screens/SobreNos";
-import Home from "../components/screens/Home";
-import CardProd from '../components/screens/CardProd';
+import DrawerNavigation from "./DrawerNavigation.js";
 
-const Stack = createNativeStackNavigator();
-
-export default function Routes () {
-    return(
-        
-        <Stack.Navigator>
-            <Stack.Screen  
-                name="Welcome"
-                component={Welcome}
-                options={{ headerShown: false}}
-            />
-            <Stack.Screen  
-                name="Login"
-                component={Login}
-                options={{ headerShown: false}}
-            />
-            <Stack.Screen  
-                name="RecuperarSenha"
-                component={RecuperarSenha}
-                options={{ headerShown: false}}
-            />
-            <Stack.Screen  
-                name="SobreNos"
-                component={SobreNos}
-                options={{ headerShown: false}}
-            />
-            <Stack.Screen  
-                name="Home"
-                component={Home}
-                options={{ headerShown: false}}
-            />
-            <Stack.Screen  
-                name="CardProd"
-                component={CardProd}
-                options={{ headerShown: false}}
-            />
-        </Stack.Navigator>
+export default function Routes() {
+    return (
+            <DrawerNavigation />
     )
 }
